+++ skbio/tree/__init__.py
@@ -26,6 +26,15 @@ Phylogenetic reconstruction
    :toctree: generated/
 
     nj
+    gme
+
+
+Tree Rearrangement
+------------------
+
+.. autosummary::
+   :toctree: generated/
+
     nni
 
 
@@ -187,7 +196,9 @@ pairwise tip-to-tip distances between trees:
 # ----------------------------------------------------------------------------
 
 from ._tree import TreeNode
-from ._nj import nj, nni
+from ._nj import nj
+from ._gme import gme
+from ._nni import nni
 from ._majority_rule import majority_rule
 from ._exception import (
     TreeError,
@@ -200,6 +211,8 @@ from ._exception import (
 __all__ = [
     "TreeNode",
     "nj",
+    "gme",
+    "nni",
     "majority_rule",
     "TreeError",
     "NoLengthError",